[refactoring] domain layer도 pacakage 분리
