+++ src/main/kotlin/com/seokjoo/todo/domain/repository/category/CategoryRepository.kt
@@ -1,6 +1,6 @@
-package com.seokjoo.todo.domain.repository
+package com.seokjoo.todo.domain.repository.category
 
-import com.seokjoo.todo.domain.entity.Category
+import com.seokjoo.todo.domain.entity.category.Category
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
