+++ src/main/kotlin/com/seokjoo/todo/domain/repository/todo/TodoRepository.kt
@@ -1,6 +1,6 @@
-package com.seokjoo.todo.domain.repository
+package com.seokjoo.todo.domain.repository.todo
 
-import com.seokjoo.todo.domain.entity.Todo
+import com.seokjoo.todo.domain.entity.todo.Todo
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
