+++ src/main/kotlin/com/seokjoo/todo/presentation/category/service/CategoryService.kt
@@ -1,6 +1,6 @@
 package com.seokjoo.todo.presentation.category.service
 
-import com.seokjoo.todo.domain.repository.CategoryRepository
+import com.seokjoo.todo.domain.repository.category.CategoryRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
